azure_function_stock_scraper/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
greenlet==1.1.3.post0
idna==3.4
lxml==4.9.1
numpy==1.23.4
pandas==1.5.1
psycopg==3.1.3
psycopg2-binary==2.9.4
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
typing-extensions==4.4.0
urllib3==1.26.12